pyproject.toml
[tool.poetry]
name = "lin655"
version = "0.1.0"
description = ""
authors = ["Adil Soubki <adil@soubki.com>"]

[tool.poetry.dependencies]
python = "^3.8"
jupyterlab = "^3.2.4"
matplotlib = "^3.5.0"
pandas = "^1.3.4"
nltk = "^3.6.5"
spacy = "^3.2.1"
conllu = "^4.4.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
